reworked mocks
