+++ src/mock/videos.js
@@ -1,0 +1,54 @@
+const videos = [
+  {
+    id: 'WQeoO7MI0Bs',
+    title: 'LEARN OPENCV in 3 HOURS with Python | Including 3x Example Projects (2020)',
+    url: 'https://www.youtube.com/embed/WQeoO7MI0Bs',
+    thumbnail:
+      'https://i.ytimg.com/vi/WQeoO7MI0Bs/hqdefault.jpg?sqp=-oaymwEYCKgBEF5IVfKriqkDCwgBFQAAiEIYAXAB&rs=AOn4CLAGDDgbGughpBZroPFgIdL4Lqat2g',
+  },
+  {
+    id: 'cbuZfY2S2UQ',
+    title: '[ 𝑷𝒍𝒂𝒚𝒍𝒊𝒔𝒕 ] 코딩할때 듣기 좋은 노래 • lofi type beat • 3 hours',
+    url: 'https://www.youtube.com/embed/cbuZfY2S2UQ',
+    thumbnail:
+      'https://i.ytimg.com/vi/cbuZfY2S2UQ/hqdefault.jpg?sqp=-oaymwEYCKgBEF5IVfKriqkDCwgBFQAAiEIYAXAB&rs=AOn4CLCB5wWjUoXrm5YaMdfnAS5hoJr9cQ',
+  },
+  {
+    id: 'blnJvW3EhWA',
+    title: 'Touch of Katana ⛩️ Japanese Lofi Hip-Hop',
+    url: 'https://www.youtube.com/embed/blnJvW3EhWA',
+    thumbnail:
+      'https://i.ytimg.com/vi/blnJvW3EhWA/hqdefault.jpg?sqp=-oaymwEZCNACELwBSFXyq4qpAwsIARUAAIhCGAFwAQ==&rs=AOn4CLDkXhff3y7sZvWPE4q7dDwmwE2VHQ',
+  },
+  {
+    id: 'rAy_3SIqT-E',
+    title: 'TypeScript Crash Course',
+    url: 'https://www.youtube.com/embed/rAy_3SIqT-E',
+    thumbnail:
+      'https://i.ytimg.com/vi/rAy_3SIqT-E/hq720.jpg?sqp=-oaymwEZCNAFEJQDSFXyq4qpAwsIARUAAIhCGAFwAQ==&rs=AOn4CLAki1P_A6fgU9YYweb4IZfB0APZ6g',
+  },
+  {
+    id: 'Sm0wwmEwqpI',
+    title: 'Python tricks: Writing user-friendly code with PEP-8 naming conventions',
+    url: 'https://www.youtube.com/embed/Sm0wwmEwqpI',
+    thumbnail:
+      'https://i.ytimg.com/vi/Sm0wwmEwqpI/hq720.jpg?sqp=-oaymwEZCNAFEJQDSFXyq4qpAwsIARUAAIhCGAFwAQ==&rs=AOn4CLB9O21HCQGRssoQV_hpzjIw6wtLlg',
+  },
+  {
+    id: 'Dy9ojlR72TM',
+    title: 'dreamy ~ lofi hip hop mix',
+    url: 'https://www.youtube.com/embed/Dy9ojlR72TM',
+    thumbnail:
+      'https://i.ytimg.com/vi/Dy9ojlR72TM/hq720.jpg?sqp=-oaymwEZCNAFEJQDSFXyq4qpAwsIARUAAIhCGAFwAQ==&rs=AOn4CLCMKB6w4wxtWTBztp7Akfpz8ifNfA',
+  },
+  {
+    id: 'CyrgzzxQPzk',
+    title:
+      "Easy & Powerful Arduino Alternative? #4 Pyboard (MicroPython) Beginner's Guide",
+    url: 'https://www.youtube.com/embed/CyrgzzxQPzk',
+    thumbnail:
+      'https://i.ytimg.com/vi/CyrgzzxQPzk/hq720.jpg?sqp=-oaymwEZCNAFEJQDSFXyq4qpAwsIARUAAIhCGAFwAQ==&rs=AOn4CLB_aL4gt8xGKVxTzbfcKLY0CgXJFA',
+  },
+];
+
+export default videos;
