+++ src/pages/Home/Home.page.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import AppBar from '../../components/AppBar';
 import VideoGrid from '../../components/VideoGrid';
-import { videos } from '../../mock';
+import videos from '../../mock';
 
 export default function HomePage() {
   return (
