+++ src/pages/Watch/Watch.page.jsx
@@ -3,7 +3,7 @@ import { useParams } from 'react-router-dom';
 import { Grid } from '@material-ui/core';
 import AppBar from '../../components/AppBar';
 import VideoDisplay from '../../components/VideoDisplay';
-import { videos } from '../../mock';
+import videos from '../../mock';
 import '../../global.css';
 
 export default function WatchVideoPage() {
@@ -18,10 +18,10 @@ export default function WatchVideoPage() {
     <>
       <AppBar />
       <Grid container className="content">
-        <Grid item style={{ backgroundColor: 'tomato' }} xs={12} sm={8} md={10}>
+        <Grid item xs={12} sm={8} md={10}>
           <VideoDisplay video={video} />
         </Grid>
-        <Grid item style={{ backgroundColor: '#cfe8fc' }} sm>
+        <Grid item sm>
           &nbsp;
         </Grid>
       </Grid>
